+++ CHANGELOG.md
@@ -18,7 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Orion`: Extended event sourcing pattern to connection types
   - Made Event.js diffing engine able to process ObjectId diffing
   - Built out class-based event sourcing for `PathwaysPersonConnection` CUD ops against pulse-core [#779](https://github.com/pulse-data/wave-db-mgmt/pull/779)
-  
+- `Orion` The custom regional breakdown export is back and supports multiple breakdowns and data sources [#775](https://github.com/pulse-data/wave-db-mgmt/pull/775)
+
 ## [3.7.0] - Started: 9-11-20
 
 ### Added